Move processCallbacks to the generic data provider
