+++ src/Plugin/resource/DataProvider/DataProvider.php
@@ -12,6 +12,7 @@ use Drupal\restful\Exception\ServiceUnavailableException;
 use Drupal\restful\Http\RequestInterface;
 use Drupal\restful\Plugin\resource\Field\ResourceFieldCollectionInterface;
 use Drupal\restful\Plugin\resource\Field\ResourceFieldInterface;
+use Drupal\restful\Resource\ResourceManager;
 
 abstract class DataProvider implements DataProviderInterface {
 
@@ -429,6 +430,30 @@ abstract class DataProvider implements DataProviderInterface {
   protected static function getLanguage() {
     // Move to its own method to allow unit testing.
     return $GLOBALS['language']->language;
+  }
+
+  /**
+   * Applies the process callbacks.
+   *
+   * @param mixed $value
+   *   The value for the field.
+   * @param ResourceFieldInterface $resource_field
+   *   The resource field.
+   *
+   * @return mixed
+   *   The value after applying all the process callbacks.
+   *
+   * @throws \Drupal\restful\Exception\ServerConfigurationException
+   */
+  protected function processCallbacks($value, ResourceFieldInterface $resource_field) {
+    $process_callbacks = $resource_field->getProcessCallbacks();
+    if (!$value || empty($process_callbacks)) {
+      return $value;
+    }
+    foreach ($process_callbacks as $process_callback) {
+      $value = ResourceManager::executeCallback($process_callback, array($value));
+    }
+    return $value;
   }
 
 }
